Made the methods private in auditcommand and extracted them into moderationutils
