+++ application/src/main/java/org/togetherjava/tjbot/features/moderation/ModerationUtils.java
@@ -1,6 +1,7 @@
 package org.togetherjava.tjbot.features.moderation;
 
 import net.dv8tion.jda.api.EmbedBuilder;
+import net.dv8tion.jda.api.JDA;
 import net.dv8tion.jda.api.Permission;
 import net.dv8tion.jda.api.entities.Guild;
 import net.dv8tion.jda.api.entities.IPermissionHolder;
@@ -25,11 +26,17 @@ import javax.annotation.Nullable;
 
 import java.awt.Color;
 import java.time.Instant;
+import java.time.ZoneOffset;
 import java.time.temporal.ChronoUnit;
 import java.time.temporal.TemporalUnit;
+import java.util.ArrayList;
+import java.util.Collection;
+import java.util.List;
+import java.util.Map;
 import java.util.Optional;
 import java.util.function.Predicate;
 import java.util.regex.Pattern;
+import java.util.stream.Collectors;
 
 /**
  * Utility class offering helpers revolving around user moderation, such as banning or kicking.
@@ -273,7 +280,7 @@ public class ModerationUtils {
      * Creates a message to be displayed as response to a moderation action.
      * <p>
      * Essentially, it informs others about the action, such as "John banned Bob for playing with
-     * the fire.".
+     * the fire".
      *
      * @param author the author executing the action
      * @param action the action that is executed
@@ -442,4 +449,84 @@ public class ModerationUtils {
      */
     record TemporaryData(Instant expiresAt, String duration) {
     }
+
+    /**
+     * Splits a list of moderation records into discrete pages capped at 10 items each.
+     *
+     * @param actions the list of chronological actions against a target
+     * @return a list of sub-lists where each sub-list contains a maximum of 10 items
+     */
+    public static List<List<ActionRecord>> groupActionsByPages(List<ActionRecord> actions) {
+        List<List<ActionRecord>> groupedActions = new ArrayList<>();
+        final int maxPageLength = 10;
+
+        for (int i = 0; i < actions.size(); i++) {
+            if (i % maxPageLength == 0) {
+                groupedActions.add(new ArrayList<>(maxPageLength));
+            }
+            groupedActions.getLast().add(actions.get(i));
+        }
+
+        return groupedActions;
+    }
+
+    /**
+     * Generates a structural text overview outlining the count total of each action type.
+     *
+     * @param actions a collection of history records
+     * @return a formatted markdown description summary
+     */
+    public static String createSummaryMessageDescription(Collection<ActionRecord> actions) {
+        int actionAmount = actions.size();
+
+        String shortSummary = "There are **%s actions** against the user."
+            .formatted(actionAmount == 0 ? "no" : actionAmount);
+
+        if (actionAmount == 0) {
+            return shortSummary;
+        }
+
+        Map<ModerationAction, Long> actionTypeToCount = actions.stream()
+            .collect(Collectors.groupingBy(ActionRecord::actionType, Collectors.counting()));
+
+        String typeCountSummary = actionTypeToCount.entrySet()
+            .stream()
+            .filter(typeAndCount -> typeAndCount.getValue() > 0)
+            .sorted(Map.Entry.<ModerationAction, Long>comparingByValue().reversed())
+            .map(typeAndCount -> "- **%s**: %d".formatted(typeAndCount.getKey(),
+                    typeAndCount.getValue()))
+            .collect(Collectors.joining("\n"));
+
+        return shortSummary + "\n" + typeCountSummary;
+    }
+
+    /**
+     * Converts an action record item asynchronously into a formatted embed data field.
+     *
+     * @param action the record data item
+     * @param jda the active JDA instance used to resolve the moderator's handle
+     * @return a field representation task mapping out the execution card detail
+     */
+    public static RestAction<MessageEmbed.Field> actionToField(ActionRecord action, JDA jda) {
+        return jda.retrieveUserById(action.authorId())
+            .map(author -> author == null ? "(unknown user)" : author.getName())
+            .map(authorText -> {
+                String expiresAtFormatted = action.actionExpiresAt() == null ? ""
+                        : "\nTemporary action, expires at: " + net.dv8tion.jda.api.utils.TimeUtil
+                            .getDateTimeString(action.actionExpiresAt().atOffset(ZoneOffset.UTC));
+
+                String fieldName = "%s by %s".formatted(action.actionType().name(), authorText);
+                String fieldDescription =
+                        """
+                                %s
+                                Issued at: %s%s
+                                """.formatted(action.reason(),
+                                net.dv8tion.jda.api.utils.TimeUtil
+                                    .getDateTimeString(action.issuedAt().atOffset(ZoneOffset.UTC)),
+                                expiresAtFormatted);
+
+                return new MessageEmbed.Field(fieldName, fieldDescription, false);
+            });
+    }
+
 }
